Moved latency from float to unsigned char

+++ CalibTracker/SiStripESProducers/python/services/SiStripLatencyGeneratorService_cfi.py
@@ -3,7 +3,7 @@ import FWCore.ParameterSet.Config as cms
 SiStripLatencyGenerator = cms.Service("SiStripLatencyGenerator",
                                      # printDebug = cms.untracked.uint32(5),
                                      file = cms.FileInPath('CalibTracker/SiStripCommon/data/SiStripDetInfo.dat'),
-                                     latency = cms.double(1.),
+                                     latency = cms.uint32(1),
                                      mode = cms.uint32(37)
                                      )
